chore: rename enroll to enrol
